feat: simplify git workflow by only applying it to files with partially staged changes
This change simplifies and speeds up the git workflow by restoring the backup stash before running tasks, and only doing diff/apply to files with both staged and unstaged changes. It also means that a file with only unstaged changes will be present on disk while the tasks are running, so for example unstaged edits to eslint rules will still apply.  This can be considered a fix because the behaviour changed from v9->v10, and this restores the previous behaviour.